requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
et-xmlfile==1.1.0
fastapi==0.104.1
h11==0.14.0
httptools==0.6.1
idna==3.4
kombu==5.3.3
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
prompt-toolkit==3.0.39
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.9
websockets==12.0
icecream==2.1.3
flower==2.0.1